requirements.txt
annotated-types==0.7.0
anyio==4.11.0
autogen-agentchat==0.7.5
autogen-core==0.7.5
autogen-ext==0.7.5
blinker==1.9.0
certifi==2025.10.5
click==8.3.0
distro==1.9.0
Flask==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonref==1.1.0
MarkupSafe==3.0.3
numpy==2.3.4
nvidia-nccl-cu12==2.28.3
openai==2.6.0
opentelemetry-api==1.38.0
pandas==2.3.3
pillow==12.0.0
protobuf==5.29.5
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
xgboost==3.1.0
zipp==3.23.0
